# back/src/main/java/com/producttrialmaster/back/dto/CartDTO.java
package com.producttrialmaster.back.dto;

import java.util.List;

public class CartDTO {
    
    private List<CartItemDTO> items;

    public CartDTO() {}

    public CartDTO(List<CartItemDTO> items) {
        this.items = items;
    }

    public List<CartItemDTO> getItems() {
        return items;
    }

    public void setItems(List<CartItemDTO> items) {
        this.items = items;
    }
    
}