# back/src/main/java/com/producttrialmaster/back/dto/CartItemUpdateQuantityDTO.java
package com.producttrialmaster.back.dto;

public class CartItemUpdateQuantityDTO {
    
    private int quantity;

    public int getQuantity() {
        return quantity;
    }

    public void setQuantity(int quantity) {
        this.quantity = quantity;
    }
}